Fix: multiple servers announcing themselves with the same invite-code
A user can be silly and use the same configuration for two servers
on different ports.
To deal with the situation, disconnect the old server and use the
new server. Most likely the old server reconnects, kicks off the
new server, and this battle continues till the user realises
what is going on.
Not much more we can do about it.