Codechange: rework how connections are setup
Instead of using a magic queue and wait for connection methods
to arrive on it, make a bit more use of coroutines and just write
in a single functions all the connection methods and their order.
This makes the whole function a lot less magic and more readable.

Accidental this might fix several issues, and report better
statistics about what is failing when.