Fix: one client could attempt multiple simultaneous connections to the same server (#52)
Although one could consider edge-cases where this can be useful
(refreshing a server while also joining it, for example), it also
allows for abuse. For example, if you hit Refresh like a maniac,
the client tries to establish many connections to the same server.

So instead of allowing this, simply drop all previous attempts
if the same client tries to connect to the exact same server.

The edge-case is not worth fixing, as if you join a server, the
refresh information is not useful to you anymore anyway.