Add: use enable_strong_referenced_tasks() to prevent destroyed-while-pending (#164)
In many places in the code we use asyncio.create_task() in a
fire&forget way, where we don't store the result value anywhere.

Sadly, this can lead to the GC removing the task before it is done
executing. It doesn't seem to happen that often, but some traces
on Sentry suggest it might be the cause of some of the trouble.