Fix: don't remove a server twice when a connection is quick enough
When a server registers for a second time, the first server is
removed. But this also causes the connection to be closed. It
was possible for the code to race, where the forcefully removing
of the server happened and the closing connection caused a server
to be removed. In rare cases this could cause the new server
to be removed too.